requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
python-dotenv
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
